[proliferate] fix plugin bundle ci checks
