+++ server/proliferate/server/cloud/plugins/catalog/first_party.py
@@ -1,13 +1,12 @@
 from __future__ import annotations
 
-from dataclasses import asdict, replace
 import hashlib
 import json
+from dataclasses import asdict, replace
 
 from proliferate.server.cloud.mcp_catalog.domain.types import CatalogEntry
 from proliferate.server.cloud.plugins.catalog.domain.types import PluginPackage, PluginSkill
 from proliferate.server.cloud.plugins.catalog.provenance import adapted_skill
-
 
 OPENAI_PLUGINS_REPO = "https://github.com/openai/plugins"
 OPENAI_PLUGINS_SOURCE_REF = "7955f1db081ddb3e14387b27cd65cf96b3e33931"
@@ -75,7 +74,9 @@ _FIRST_PARTY_SKILLS_BY_CATALOG_ENTRY_ID: dict[str, tuple[PluginSkill, ...]] = {
             catalog_entry_id="github",
             id="triage",
             display_name="GitHub triage",
-            description="Inspect repositories, issues, pull requests, and checks before taking action.",
+            description=(
+                "Inspect repositories, issues, pull requests, and checks before taking action."
+            ),
             relative_path="github/triage.md",
             source_path="plugins/github/skills/github/SKILL.md",
             source_sha256="81dbdd90934fe86a79ddc4790fd211e5fca866302a74090ad153395f56f2bd42",
@@ -86,7 +87,9 @@ _FIRST_PARTY_SKILLS_BY_CATALOG_ENTRY_ID: dict[str, tuple[PluginSkill, ...]] = {
             catalog_entry_id="github",
             id="address-comments",
             display_name="Address PR comments",
-            description="Review unresolved PR feedback and plan changes against the current branch.",
+            description=(
+                "Review unresolved PR feedback and plan changes against the current branch."
+            ),
             relative_path="github/address-comments.md",
             source_path="plugins/github/skills/gh-address-comments/SKILL.md",
             source_sha256="c1ebc337357402f7faabafe712e0c463981a65f736453efe52abd305bcb74769",
@@ -113,7 +116,9 @@ _FIRST_PARTY_SKILLS_BY_CATALOG_ENTRY_ID: dict[str, tuple[PluginSkill, ...]] = {
             source_path="plugins/linear/skills/linear/SKILL.md",
             source_sha256="8a409240f8f9f70363550340b66ff31377b638d25e780e081546248ed417c6dc",
             source_license="MIT",
-            notes="Adapted to read-first behavior; mutation guidance requires explicit user intent.",
+            notes=(
+                "Adapted to read-first behavior; mutation guidance requires explicit user intent."
+            ),
         ),
     ),
     "slack": (
@@ -132,7 +137,9 @@ _FIRST_PARTY_SKILLS_BY_CATALOG_ENTRY_ID: dict[str, tuple[PluginSkill, ...]] = {
             catalog_entry_id="slack",
             id="channel-summary",
             display_name="Slack channel summary",
-            description="Summarize channel or thread activity with citations to the source messages.",
+            description=(
+                "Summarize channel or thread activity with citations to the source messages."
+            ),
             relative_path="slack/channel-summary.md",
             source_path="plugins/slack/skills/slack-channel-summarization/SKILL.md",
             source_sha256="e4c87663ca8c12e3df530963b5f1e6669e9b8e2ec551f02152fb8f410bdb92de",
@@ -164,7 +171,9 @@ _FIRST_PARTY_SKILLS_BY_CATALOG_ENTRY_ID: dict[str, tuple[PluginSkill, ...]] = {
             catalog_entry_id="notion",
             id="research-documentation",
             display_name="Notion research documentation",
-            description="Organize research findings into Notion pages with source-backed sections.",
+            description=(
+                "Organize research findings into Notion pages with source-backed sections."
+            ),
             relative_path="notion/research-documentation.md",
             source_path="plugins/notion/skills/notion-research-documentation/SKILL.md",
             source_sha256="0e6377705bcac9166a1df6bb96b1f1417d15f89be561b80a421d9208f50d1bad",
@@ -174,7 +183,9 @@ _FIRST_PARTY_SKILLS_BY_CATALOG_ENTRY_ID: dict[str, tuple[PluginSkill, ...]] = {
             catalog_entry_id="notion",
             id="spec-to-implementation",
             display_name="Notion spec to implementation",
-            description="Use Notion specs as implementation context and keep derived tasks traceable.",
+            description=(
+                "Use Notion specs as implementation context and keep derived tasks traceable."
+            ),
             relative_path="notion/spec-to-implementation.md",
             source_path="plugins/notion/skills/notion-spec-to-implementation/SKILL.md",
             source_sha256="4d60e6d384ca631f13b19d83dedde864f4eb92b12c6955d54ad434b0f0a7011a",
@@ -197,7 +208,9 @@ _FIRST_PARTY_SKILLS_BY_CATALOG_ENTRY_ID: dict[str, tuple[PluginSkill, ...]] = {
             catalog_entry_id="supabase",
             id="postgres-best-practices",
             display_name="Supabase Postgres best practices",
-            description="Review schema, query, and migration work against Supabase Postgres practices.",
+            description=(
+                "Review schema, query, and migration work against Supabase Postgres practices."
+            ),
             relative_path="supabase/postgres-best-practices.md",
             source_path="plugins/supabase/skills/supabase-postgres-best-practices/SKILL.md",
             source_sha256="ccd6e4596bd51cf344fe76c464867c541ccc16b6d90ae7a9db449fb17588613b",
@@ -209,7 +222,9 @@ _FIRST_PARTY_SKILLS_BY_CATALOG_ENTRY_ID: dict[str, tuple[PluginSkill, ...]] = {
             catalog_entry_id="render",
             id="debug",
             display_name="Render debug",
-            description="Inspect Render service state, events, logs, and likely deployment issues.",
+            description=(
+                "Inspect Render service state, events, logs, and likely deployment issues."
+            ),
             relative_path="render/debug.md",
             source_path="plugins/render/skills/render-debug/SKILL.md",
             source_sha256="72268428eec2917201c224f3c1045369ea3575b41b94ff0c896a066f8b643c43",
@@ -220,7 +235,9 @@ _FIRST_PARTY_SKILLS_BY_CATALOG_ENTRY_ID: dict[str, tuple[PluginSkill, ...]] = {
             catalog_entry_id="render",
             id="monitor",
             display_name="Render monitor",
-            description="Summarize Render service health and recent incidents from available data.",
+            description=(
+                "Summarize Render service health and recent incidents from available data."
+            ),
             relative_path="render/monitor.md",
             source_path="plugins/render/skills/render-monitor/SKILL.md",
             source_sha256="b8ca61fb72d95bde1ce438939b14310a4b6cebebddcc2e74cdc6ae9cd93a0f8f",
@@ -254,7 +271,9 @@ _FIRST_PARTY_SKILLS_BY_CATALOG_ENTRY_ID: dict[str, tuple[PluginSkill, ...]] = {
             catalog_entry_id="cloudflare_docs",
             id="docs-research",
             display_name="Cloudflare docs research",
-            description="Use Cloudflare Docs as retrieval context without issuing platform mutations.",
+            description=(
+                "Use Cloudflare Docs as retrieval context without issuing platform mutations."
+            ),
             relative_path="cloudflare/docs-research.md",
             source_path="plugins/cloudflare/skills/cloudflare/SKILL.md",
             source_sha256="613c4c57c19bd6385a870b4d441e00a0d88cc6c348c447f8a88889d34b17b2d0",
@@ -267,7 +286,9 @@ _FIRST_PARTY_SKILLS_BY_CATALOG_ENTRY_ID: dict[str, tuple[PluginSkill, ...]] = {
             catalog_entry_id="neon",
             id="readonly-inspection",
             display_name="Neon read-only inspection",
-            description="Inspect Neon Postgres context using the hosted read-only MCP assumptions.",
+            description=(
+                "Inspect Neon Postgres context using the hosted read-only MCP assumptions."
+            ),
             relative_path="neon/readonly-inspection.md",
             source_path="plugins/neon-postgres/skills/neon-postgres/SKILL.md",
             source_sha256="102b4b2d25ab50aa8355f06792ee7828655af05a55be419744099b1dc8dc5fa3",
@@ -285,7 +306,10 @@ _FIRST_PARTY_SKILLS_BY_CATALOG_ENTRY_ID: dict[str, tuple[PluginSkill, ...]] = {
             source_path="plugins/gmail/skills/gmail/SKILL.md",
             source_sha256="d3eee36dc69ba2fb8388018df122d5f53e4a807c373067edad8d3df1c588f6e4",
             source_license="MIT",
-            notes="Send, archive, delete, and label operations removed because Proliferate launches Gmail as read-only.",
+            notes=(
+                "Send, archive, delete, and label operations removed because Proliferate "
+                "launches Gmail as read-only."
+            ),
         ),
         _skill(
             catalog_entry_id="gmail",
